server/pyproject.toml
[project]
name = "pulseboard-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.121.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "razorpay>=2.0.0",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
